backend/requirements.txt
sentence-transformers
langchain
python-dotenv
langchain-google-genai
ipywidgets
pypdf
qdrant_client
fastapi
uvicorn[standard]